refactor: replace preAlignmentMetadataHandler with notSupportedMetadataHandler and add baseMetadataHandler
